Backend/requirements.txt
Django==5.2
djangorestframework==3.16.1
django-cors-headers==4.9.0
djangorestframework_simplejwt==5.5.1

pdfplumber==0.11.4
python-docx==1.1.2

spacy==3.7.5
pydantic==2.7.1

numpy==2.4.3
pillow==12.1.1
python-dotenv==1.2.2

gunicorn==25.1.0